added integration test for notification
